requirements.txt
flask
flask-restful
opencv-python-headless
minio
paho-mqtt